requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
clipboard==0.0.4
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.23.6
pydantic==2.7.1
pydantic_core==2.18.2
pyperclip==1.8.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
